pyproject.toml
[project]
name = "elasticsearch7-mcp-server"
version = "1.0.0"
description = "MCP Server for interacting with Elasticsearch 7.x"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elasticsearch>=7.0.0,<8.0.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.4.0",
]

[project.license]
file = "LICENSE"

[project.scripts]
elasticsearch7-mcp-server = "elasticsearch7_mcp_server:main"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build" 